requirements.txt
black==24.1.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
idna==3.6
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
python-dotenv==1.0.1
requests==2.31.0
soupsieve==2.5
tomli==2.0.1
typing-extensions==4.9.0
urllib3==2.2.0
